requirements.txt
numpy==1.21.5
pandas==1.5.1
scikit-learn==1.2.1
tqdm==4.64.1
Pillow==9.2.0
nltk==3.7
jiwer==2.5.1
omegaconf==2.2.3
ipython==7.31.1
ipykernel==6.15.2
tensorboard==2.11.0
git+https://github.com/openai/CLIP.git